Format code with `ktfmt` and sort strings with `xsltproc`

+++ app/src/main/java/dev/itsvic/parceltracker/db/Helpers.kt
@@ -3,10 +3,9 @@ package dev.itsvic.parceltracker.db
 import dev.itsvic.parceltracker.ParcelApplication
 
 suspend fun deleteParcel(parcel: Parcel) {
-    val db = ParcelApplication.db
-    val parcelId = ParcelId(parcel.id)
-    db.parcelDao().delete(parcel)
-    db.parcelStatusDao().deleteByParcelId(parcelId)
-    if (parcel.isArchived)
-        db.parcelHistoryDao().deleteByParcelId(parcelId)
+  val db = ParcelApplication.db
+  val parcelId = ParcelId(parcel.id)
+  db.parcelDao().delete(parcel)
+  db.parcelStatusDao().deleteByParcelId(parcelId)
+  if (parcel.isArchived) db.parcelHistoryDao().deleteByParcelId(parcelId)
 }
